correcciones

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: pcabanas <pcabanas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 17:04:18 by pcabanas          #+#    #+#             */
-/*   Updated: 2024/01/31 16:41:30 by pcabanas         ###   ########.fr       */
+/*   Updated: 2024/02/08 15:41:26 by pcabanas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 //calculate the length of a string
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len] != '\0')
